feat: split `@pkgr/utils` into `@pkgr/core` and `@pkgr/browser`
close #352
