+++ packages/browser/openChrome.applescript
