+++ packages/browser/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@pkgr/utils",
-  "version": "2.4.2",
+  "name": "@pkgr/browser",
+  "version": "0.0.0",
   "type": "module",
-  "description": "Shared utils for `@pkgr` packages or any package else",
+  "description": "Shared browser utils for `@pkgr` packages or any package else",
   "repository": "git+https://github.com/un-ts/pkgr.git",
-  "homepage": "https://github.com/un-ts/pkgr/blob/master/packages/utils",
+  "homepage": "https://github.com/un-ts/pkgr/blob/master/packages/browser",
   "author": "JounQin (https://www.1stG.me) <admin@1stg.me>",
   "funding": "https://opencollective.com/unts",
   "license": "MIT",
@@ -20,12 +20,14 @@
   },
   "types": "./lib/index.d.ts",
   "files": [
-    "lib"
+    "lib",
+    "openChrome.applescript"
   ],
   "dependencies": {
-    "@pkgr/core": "^0.0.0",
-    "fast-glob": "^3.3.1",
-    "is-glob": "^4.0.3"
+    "cross-spawn": "^7.0.3",
+    "open": "^9.1.0",
+    "picocolors": "^1.0.0",
+    "tslib": "^2.6.1"
   },
   "publishConfig": {
     "access": "public"
