+++ packages/core/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@pkgr/utils",
-  "version": "2.4.2",
+  "name": "@pkgr/core",
+  "version": "0.0.0",
   "type": "module",
-  "description": "Shared utils for `@pkgr` packages or any package else",
+  "description": "Shared core module for `@pkgr` packages or any package else",
   "repository": "git+https://github.com/un-ts/pkgr.git",
-  "homepage": "https://github.com/un-ts/pkgr/blob/master/packages/utils",
+  "homepage": "https://github.com/un-ts/pkgr/blob/master/packages/core",
   "author": "JounQin (https://www.1stG.me) <admin@1stg.me>",
   "funding": "https://opencollective.com/unts",
   "license": "MIT",
@@ -22,11 +22,6 @@
   "files": [
     "lib"
   ],
-  "dependencies": {
-    "@pkgr/core": "^0.0.0",
-    "fast-glob": "^3.3.1",
-    "is-glob": "^4.0.3"
-  },
   "publishConfig": {
     "access": "public"
   },
