+++ packages/core/src/index.ts
@@ -1,4 +1,2 @@
 export * from './constants.js'
 export * from './helpers.js'
-export * from './monorepo.js'
-export * from '@pkgr/core'
