+++ packages/utils/src/constants.ts
@@ -1,5 +1,3 @@
-import { createRequire } from 'node:module'
-
 export const DEV = 'development' as const
 export const PROD = 'production' as const
 
@@ -9,14 +7,6 @@ export const __DEV__ = NODE_ENV === DEV
 export const __PROD__ = NODE_ENV === PROD
 
 export const NODE_MODULES_REG = /[/\\]node_modules[/\\]/
-
-export const CWD = process.cwd()
-
-export const cjsRequire =
-  typeof require === 'undefined' ? createRequire(import.meta.url) : require
-
-// eslint-disable-next-line n/no-deprecated-api, sonar/deprecation
-export const EXTENSIONS = ['.ts', '.tsx', ...Object.keys(cjsRequire.extensions)]
 
 export const SCRIPT_RUNNERS = {
   npm: 'npx',
